gitpod_url

+++ utils/util_ybtserver_metrics.sql
@@ -16,7 +16,6 @@ drop function if exists fn_yb_tserver_metrics_snap;
 drop view if exists vw_yb_tserver_metrics_snapshot_tablets;
 drop view if exists vw_yb_tserver_metrics_report;
 drop view if exists vw_yb_tserver_metrics_snapshot_tablets_metrics;
-drop view if exists vw_yb_tserver_metrics_snap_and_show_tablet_load;
 drop function if exists fn_yb_tserver_metrics_snap_and_show_tablet_load_ct;
 drop view if exists vw_yb_tserver_metrics_snap_and_show_tablet_load;
 drop function if exists fn_yb_tserver_metrics_snap_table;
@@ -36,13 +35,16 @@ create or replace function fn_yb_tserver_metrics_snap(snaps_to_keep int default 
 returns timestamptz as $DO$
 declare i record; 
 begin
-    delete from tbl_yb_tserver_metrics_snapshots 
-    where 1=1
-    and ts not in (
-        select distinct ts          
-        from tbl_yb_tserver_metrics_snapshots
-        order by ts desc
-        limit snaps_to_keep);
+
+    if (select count(*) from tbl_yb_tserver_metrics_snapshots ) > 0 then
+        delete from tbl_yb_tserver_metrics_snapshots 
+        where 1=1
+        and ts not in (
+            select distinct ts          
+            from tbl_yb_tserver_metrics_snapshots
+            order by ts desc
+            limit snaps_to_keep);
+    end if;
 
     for i in (select host from yb_servers()) loop 
          execute format('DROP TABLE if exists tbl_temp');
@@ -169,12 +171,13 @@ where 1=1
 and table_name not in ('tbl_yb_tserver_metrics_snapshots')
 order by ts desc, namespace_name, table_name, table_id, host, tablet_id, is_raft_leader, "table" desc, value desc, metric_name;
 
+
 -- select * from vw_yb_tserver_metrics_snap_and_show_tablet_load;
 
 
 
--- crosstab view
-
+-- crosstab view, not being used, replace with function
+/*
 create or replace view vw_yb_tserver_metrics_snap_and_show_tablet_load_ct as 
 select 
    ct_row_name
@@ -193,11 +196,12 @@ select
         order by 1, 2 desc,3
         $$) 
      as (ct_row_name text, "ct_rocksdb_number_db_seek" numeric, "ct_rocksdb_number_db_next" numeric, "ct_rows_inserted" decimal);
+*/
 
 
 -- crosstab function
 
-create or replace function fn_yb_tserver_metrics_snap_and_show_tablet_load_ct(isLocal int default 0)
+create or replace function fn_yb_tserver_metrics_snap_and_show_tablet_load_ct(gitpod_url text default '127.0.0.1')
 returns table (
     row_name text, 
     rocksdb_number_db_seek numeric,
@@ -208,7 +212,7 @@ language plpgsql
 as $DO$
 begin
 
-    if isLocal = 1 then
+    if gitpod_url = '127.0.0.1' then
         return query
         select 
         ct_row_name
@@ -220,7 +224,8 @@ begin
                 format('%s | %s | %s | %s | %s', namespace_name, table_name, format('http://%s:7000/table?id=%s',host,table_id),tablet_id, case is_raft_leader when 0 then ' ' else 'L' end) vw_row_name, 
                 metric_name category, 
                 sum(value)
-            from vw_yb_tserver_metrics_snap_and_show_tablet_load 
+            from vw_yb_tserver_metrics_snap_and_show_tablet_load
+            -- from fn_yb_tserver_metrics_snap_and_show_tablet_load(gitpod_url)
             where 1=1
             and metric_name in ('rocksdb_number_db_seek','rocksdb_number_db_next','rows_inserted') 
             group by namespace_name, table_name, host, table_id, tablet_id, is_raft_leader, metric_name
@@ -230,28 +235,71 @@ begin
      else
         return query
         select 
-            ct_row_name
-            , "ct_rocksdb_number_db_seek"
-            , "ct_rocksdb_number_db_next"
-            , "ct_rows_inserted"
-            from crosstab($$
-                select 
-                    format('%s | %s | %s | %s | %s', namespace_name, table_name, table_id, tablet_id, case is_raft_leader when 0 then ' ' else 'L' end) vw_row_name, 
-                    metric_name category, 
-                    sum(value)
-                from vw_yb_tserver_metrics_snap_and_show_tablet_load 
-                where 1=1
-                and metric_name in ('rocksdb_number_db_seek','rocksdb_number_db_next','rows_inserted') 
-                group by namespace_name, table_name, host, table_id, tablet_id, is_raft_leader, metric_name
-                order by 1, 2 desc,3
-                $$) 
-            as (ct_row_name text, "ct_rocksdb_number_db_seek" numeric, "ct_rocksdb_number_db_next" numeric, "ct_rows_inserted" decimal);
+        REPLACE( ct_row_name, '127.0.0.1',  gitpod_url ) as ct_row_name
+        , "ct_rocksdb_number_db_seek"
+        , "ct_rocksdb_number_db_next"
+        , "ct_rows_inserted"
+        from crosstab($$
+            select 
+                format('%s | %s | %s | %s | %s', namespace_name, table_name, format('http://%s:7000/table?id=%s',host,table_id),tablet_id, case is_raft_leader when 0 then ' ' else 'L' end) vw_row_name, 
+                metric_name category, 
+                sum(value)
+            from vw_yb_tserver_metrics_snap_and_show_tablet_load
+            -- from fn_yb_tserver_metrics_snap_and_show_tablet_load(gitpod_url)
+            where 1=1
+            and metric_name in ('rocksdb_number_db_seek','rocksdb_number_db_next','rows_inserted') 
+            group by namespace_name, table_name, host, table_id, tablet_id, is_raft_leader, metric_name
+            order by 1, 2 desc,3
+            $$) 
+        as (ct_row_name text, "ct_rocksdb_number_db_seek" numeric, "ct_rocksdb_number_db_next" numeric, "ct_rows_inserted" decimal);
      end if;
 end; $DO$;
 
 
+create or replace function fn_yb_tserver_metrics_snap_and_show_tablet_load(gitpod_url text default '127.0.0.1')
+returns table (
+    value numeric, 
+    rate numeric,
+    namespace_name text,
+    table_name text,
+    table_id text,
+    host text,
+    tablet_id text,
+    is_raft_leader text,
+    metric_name text,
+    percent_table numeric,
+    ops numeric
+) 
+language plpgsql
+as $BODY$
+begin
+    return query
+    select 
+        value
+        , rate
+        , namespace_name
+        , table_name
+        , table_id
+        , REPLACE(host, '127.0.0.1',  gitpod_url ) as host
+        , tablet_id
+        , is_raft_leader
+        , metric_name
+        , to_char(100*value/sum(value)
+            over(
+                partition by
+                namespace_name, table_name, table_id, metric_name),'999%') as "percent_table"
+        , sum(value)
+            over(
+                partition by
+                namespace_name, table_name, table_id, metric_name) as "ops"
+        from vw_yb_tserver_metrics_last , fn_yb_tserver_metrics_snap()
+        where 1=1
+        and table_name not in ('tbl_yb_tserver_metrics_snapshots')
+        order by ts desc, namespace_name, table_name, table_id, host, tablet_id, is_raft_leader, "table" desc, value desc, metric_name;
+end; $BODY$;
 
-create or replace function fn_yb_tserver_metrics_snap_table(isLocal int default 0)
+
+create or replace function fn_yb_tserver_metrics_snap_table(gitpod_url text default '127.0.0.1')
 returns table (
     row_name text, 
     rocksdb_number_db_seek numeric,
@@ -263,26 +311,33 @@ as $DO$
 begin
     return query
     select * 
-    from fn_yb_tserver_metrics_snap_and_show_tablet_load_ct(isLocal);
+    from fn_yb_tserver_metrics_snap_and_show_tablet_load_ct(gitpod_url);
 end; $DO$;
 
 
-create or replace function fn_yb_create_stmts() 
+create or replace function fn_yb_create_stmts(gitpod_url text default '127.0.0.1') 
 returns timestamptz as $DO$
 begin
 
-    if (select count(*) from pg_prepared_statements where 1=1 and name in ('stmt_util_metrics_snap_tablet','stmt_util_metrics_snap_table','stmt_util_metrics_snap_reset')) > 0  then 
+    if (select count(*) from pg_prepared_statements where 1=1 and name = 'stmt_util_metrics_snap_reset') > 0  then 
         deallocate stmt_util_metrics_snap_reset;
+    end if;
+
+    if (select count(*) from pg_prepared_statements where 1=1 and name = 'stmt_util_metrics_snap_table') > 0  then 
         deallocate stmt_util_metrics_snap_table;
+    end if;
+
+    if (select count(*) from pg_prepared_statements where 1=1 and name = 'stmt_util_metrics_snap_tablet') > 0  then 
         deallocate stmt_util_metrics_snap_tablet;
     end if;
 
     execute format('prepare stmt_util_metrics_snap_reset as select '''' as "ybwr metrics" where fn_yb_tserver_metrics_snap() is null');
-    execute format('prepare stmt_util_metrics_snap_table as select row_name as "[dbname | relname | tableid | tabletid | isLeader]", rocksdb_number_db_seek, rocksdb_number_db_next, rows_inserted from fn_yb_tserver_metrics_snap_table(0)');
-    execute format(' prepare stmt_util_metrics_snap_tablet as select * from vw_yb_tserver_metrics_snap_and_show_tablet_load where 1=1 and metric_name in (''rows_inserted'',''rocksdb_number_db_seek'',''rocksdb_number_db_next'')');
+
+    execute format('prepare stmt_util_metrics_snap_table as select row_name as "[dbname | relname | tableid | tabletid | isLeader]", rocksdb_number_db_seek, rocksdb_number_db_next, rows_inserted from fn_yb_tserver_metrics_snap_table(''%s'')',gitpod_url);
+
+    execute format(' prepare stmt_util_metrics_snap_tablet as select * from fn_yb_tserver_metrics_snap_and_show_tablet_load(''%s'') where 1=1 and metric_name in (''rows_inserted'',''rocksdb_number_db_seek'',''rocksdb_number_db_next'')',gitpod_url);
 
   return clock_timestamp(); 
 end; 
 $DO$ language plpgsql;
-
